people-detection/requirements.txt
numpy
opencv-python
opencv-contrib-python
matplotlib
Pillow
torch
torchvision
ultralytics
scikit-learn
tqdm
pandas
filterpy
scikit-image
lap
numba
seaborn
requests
flask
deep-sort-realtime